AI AGENTS/advanced-research-agent/requirements.txt
streamlit
langgraph
langchain-community
langchain-openai
langchain-nvidia-ai-endpoints
faiss-cpu
pydantic
requests
python-dotenv
typing-extensions
